[PATCH]  mm: Speed up PID map reads for Android's use case
Android frequently reads /proc/self/maps inside Bionic in order to
determine the size of the current process' stack when pthreads are used.
When the current process has several thousands of mappings though, such
as for a game, a lot of time is wasted printing out every mapping when
Bionic only cares about finding the stack mapping. We can optimize PID
map reads for Android by caching the VMA that has the process' stack and
only printing out the stack mapping every time the maps are read. This
not only massively speeds up map reads for large processes, but also
reduces the amount of time the mmap semaphore is held, which can provide other
potential benefits when there's a writer waiting.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Zlatan Radovanovic <zlatan.radovanovic@fet.ba>
Signed-off-by: Jebaitedneko <Jebaitedneko@gmail.com>
Signed-off-by: ImLonely13 <gabutuhaku@gmail.com>